# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /config:
    get:
      operationId: getConfig
      tags: [config]
      summary: Get the current configuration
      responses:
        "200":
          description: Current configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
    put:
      operationId: updateConfig
      tags: [config]
      summary: Update configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                niche:
                  type: string
                morningPostTime:
                  type: string
                afternoonPostTime:
                  type: string
                eveningPostTime:
                  type: string
                nightPostTime:
                  type: string
                lateNightPostTime:
                  type: string
                midnightPostTime:
                  type: string
                language:
                  type: string
                autoApprove:
                  type: boolean
                instagramAccountId:
                  type: string
                metaAccessToken:
                  type: string
                imageSource:
                  type: string
                  enum: [ai, search]
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
  /config/exchange-token:
    post:
      operationId: exchangeToken
      tags: [config]
      summary: Exchange short-lived Meta token for long-lived token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shortLivedToken:
                  type: string
                appId:
                  type: string
                appSecret:
                  type: string
              required:
                - shortLivedToken
      responses:
        "200":
          description: Token exchange result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenExchangeResult"
  /config/meta-login:
    post:
      operationId: metaLogin
      tags: [config]
      summary: Save Meta access token from Facebook JS SDK login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessToken:
                  type: string
              required:
                - accessToken
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenExchangeResult"
...
components:
  schemas:
...
    AppConfig:
      type: object
      properties:
        niche:
          type: string
        morningPostTime:
          type: string
        afternoonPostTime:
          type: string
        eveningPostTime:
          type: string
        nightPostTime:
          type: string
        lateNightPostTime:
          type: string
        midnightPostTime:
          type: string
        language:
          type: string
        autoApprove:
          type: boolean
        instagramAccountId:
          type: string
        metaAccessToken:
          type: string
        imageSource:
          type: string
          enum: [ai, search]
          default: ai
      required:
        - niche
        - morningPostTime
        - afternoonPostTime
        - eveningPostTime
        - nightPostTime
        - lateNightPostTime
        - midnightPostTime
        - language
        - autoApprove
        - instagramAccountId
        - metaAccessToken
        - imageSource
...
    TokenExchangeResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        instagramAccountId:
          type: string
          nullable: true
        tokenType:
          type: string
      required:
        - success
        - message